test(stats): organize lifetime summary repair tests
- Extract shared lifetime test fixtures
- Move repair scenarios into a dedicated test file
